requirements.txt
pillow
dotenv
httpx